+++ CHANGES.rst
@@ -1,6 +1,12 @@
 0.7.2 (unreleased)
 ==================
 
+photom
+------
+
+- Adds explicit test that photometric keywords are preserved for spectroscopic data. [#513]
+  
+
 0.7.1 (2022-05-19)
 ==================